// src/pages/wms/Masters/Product_Master/product-wms.types.ts
export type TProduct = {

  prin_code: string;
  prin_name?: string;
  prod_code: string;
  prod_name: string;
  group_code: string;
  group_name?: string;
  brand_code: string;
  brand_name?: string;
  packdesc?: string;
  barcode?: string;
  p_uom?: string;
  suom?: string;
  length?: number;
  breadth?: number;
  height?: number;
  volume?: number;
  gross_wt?: number;
  net_wt?: number;
  foc?: string;
  cpu?: number;
  harm_code?: string;
  imco_code?: string;
  kitting?: string;
  manu_code?: string;
  manu_name?: string;
  base_price?: number;
  flat_storage?: number;
  site_type?: string;
  site_ind?: string;
  pack_key?: string;
  prod_ti?: number;
  prod_hi?: number;
  chargetime?: string;
  prod_status?: string;
  shelf_life?: number;
  category_abc?: string;
  reord_level?: number;
  reord_qty?: number;
  alt_prod_code?: string;
  pref_site?: string;
  pref_loc_from?: string;
  pref_loc_to?: string;
  pref_aisle_from?: string;
  pref_aisle_to?: string;
  pref_col_from?: number;
  pref_col_to?: number;
  pref_ht_from?: number;
  pref_ht_to?: number;
  uppp?: number;
  chk_manucode?: string;
  chk_lotno?: string;
  chk_mfgexpdt?: string;
  puom_volume?: number;
  puom_netwt?: number;
  puom_grosswt?: number;
  l_uom?: string;
  luppp?: number;
  uom_count: number;
  prod_type?: number;
  company_code?: string;
  twoplus_uom?: string;
  upp?: number;
  wave_code?: number;
  product_stage?: string;
  co_pack?: string;
  model_number?: string;
  variant_code?: string;
  cnt_origin?: string;
  serialize?: string;
  packing?: string;
  old_upp?: number;
  avg_consumption?: number;
  prod_image_path_web?: string;
  minperiod_exppick?: number;
  rcpt_exp_limit?: number;
  qty_as_wt?: string;
  hazmat_ind?: string;
  hazmat_class?: string;
  food_ind?: string;
  pharma_ind?: string;
  special_instructions?: string;
  strength?: string;
  pack_size?: number;
  group_code_bk?: string;
  batch_type?: number;
  sap_prod_code?: string;
  sap_prod_desc?: string;
  temp_code?: string;
  edit_user?: string;
  class?: string;
  wob?: number;
  unified_code?: string;
  current_season?: string;
  product_category?: string;
  generic_article?: string;
  prod_gender?: string;
  prod_color?: string;
  prod_size?: string;
  prnt_p_code?: string;
  updated_at?: Date;
  updated_by?: string;
  created_by?: string;
  created_at?: Date;
  division?: string;
}